requirements.txt
colorama==0.4.6
discord.py==2.5.2
google-genai==1.5.0
pytz==2025.1
aiohttp==3.11.14
pillow==11.1.0
aiofiles==23.2.1
flask==3.0.3
werkzeug==3.0.3
flask-cors==4.0.1